+++ data-prep/dataprep-simple/src/main/scala/de/hpi/isg/dataprep/preparators/implementation/DefaultSplitPropertyImpl.scala
@@ -5,8 +5,11 @@ import de.hpi.isg.dataprep.components.PreparatorImpl
 import de.hpi.isg.dataprep.model.error.PreparationError
 import de.hpi.isg.dataprep.model.target.system.AbstractPreparator
 import de.hpi.isg.dataprep.preparators.define.SplitProperty
+import org.apache.spark.sql.catalyst.encoders.RowEncoder
 import org.apache.spark.sql.{Dataset, Row}
 import org.apache.spark.util.CollectionAccumulator
+
+import scala.util.Try
 
 class DefaultSplitPropertyImpl extends PreparatorImpl {
 
@@ -14,8 +17,16 @@ class DefaultSplitPropertyImpl extends PreparatorImpl {
     val preparator = abstractPreparator.asInstanceOf[SplitProperty]
     val propertyName = preparator.propertyName
     val separator = preparator.separator
+    val fromLeft = preparator.fromLeft
 
+    val rowEncoder = RowEncoder(dataFrame.schema)
     val resultDataFrame = dataFrame
+    dataFrame.flatMap(row => {
+      val x = Try{
+        row
+      }
+      x.toOption
+    })(rowEncoder)
 
     new ExecutionContext(resultDataFrame, errorAccumulator)
   }
